requirements.txt
numpy==1.17.4
matplotlib==3.1.2
jupyter==1.0.0
opencv-python==3.4.9.31
multiprocess==0.70.9
pandas==0.25.3
seaborn==0.9.0
torch==1.3.1
torchvision==0.4.2
fastai==1.0.59
pillow==6.2.1
pydub==0.23.1
scipy==1.4.1
moviepy==1.0.1
pretrainedmodels==0.7.4
librosa==0.7.1
xgboost==0.90
geffnet==0.9.5
eli5==0.10.1
lightgbm==2.3.1
scikit-learn==0.21.3
facenet_pytorch==2.2.7
albumentations==0.4.3
kornia==0.2.0

